requirements.txt
python-dotenv
pandas
pyyaml

# ccef-connections is the REQUIRED shared library for all external API connections (BigQuery,
# Action Network, Airtable, etc.). Install it from the local path — it is not on PyPI.
# Run once per machine:
#   pip install -e "C:/Users/RobKerth/OneDrive - Common Cause Education Fund/Documents/AI Interpretation/ccef-connections"
#
# Do NOT write your own API clients in this project. If a new connection type is needed,
# build it in ccef-connections and import it here.
